refactor(entities): split grouped entity files into separate files per entity
- Renamed and reorganized entity files for better clarity and maintainability.
- Each entity now has its own file following the .entity.ts naming convention.
- Removed previous grouped files like product.entities.ts, cart.entities.ts, etc.
