+++ adapters/cart/prisma/create-prisma-cart-service.adapter.ts
@@ -1,4 +1,4 @@
-import { CartEntity } from "@/domain/entities/cart.entities";
+import { CartEntity } from "@/domain/entities/cart.entity";
 import { createPrismaCartService } from "@/infra/services/cart/create-prisma-cart.service";
 import { mapDomainCartToPrismaCart } from "@/adapters/cart/prisma/map-domain-cart-to-prisma-cart";
