+++ adapters/cart/prisma/map-domain-cart-item-to-prisma-cart-item.ts
@@ -1,4 +1,4 @@
-import { CartItemEntity } from "@/domain/entities/cart.entities";
+import { CartItemEntity } from "@/domain/entities/cart-item.entity";
 import { CartItemDatabase } from "@/infra/db/types/cart";
 
 export function mapDomainCartItemToPrismaCartItem(
