+++ adapters/cart/prisma/map-domain-cart-to-prisma-cart.ts
@@ -1,4 +1,4 @@
-import { CartEntity } from "@/domain/entities/cart.entities";
+import { CartEntity } from "@/domain/entities/cart.entity";
 import { CartDatabaseInsert } from "@/infra/db/types/cart";
 import { mapDomainCartItemToPrismaCartItem } from "@/adapters/cart/prisma/map-domain-cart-item-to-prisma-cart-item";
