+++ adapters/currency/generic/create-currency-entity-or-fallback.ts
@@ -1,4 +1,4 @@
-import { CurrencyEntity } from "@/domain/entities/currency.entities";
+import { CurrencyEntity } from "@/domain/entities/currency.entity";
 import { createCurrencyEntity } from "@/adapters/currency/generic/create-currency-entity";
 
 export function createCurrencyEntityOrFallback<F>(
