+++ domain/entities/currency-value.entity.ts
@@ -1,0 +1,4 @@
+export interface CurrencyValueEntity {
+  stringValue: string;
+  numericValue: number;
+}
